Dtype preservation in preemphasis and mfcc

+++ librosa/feature/spectral.py
@@ -1685,7 +1685,7 @@ def mfcc(y=None, sr=22050, S=None, n_mfcc=20, dct_type=2, norm='ortho', lifter=0
     M = scipy.fftpack.dct(S, axis=0, type=dct_type, norm=norm)[:n_mfcc]
 
     if lifter > 0:
-        M *= 1 + (lifter / 2) * np.sin(np.pi * np.arange(1, 1 + n_mfcc) / lifter)[:, np.newaxis]
+        M *= 1 + (lifter / 2) * np.sin(np.pi * np.arange(1, 1 + n_mfcc, dtype=M.dtype) / lifter)[:, np.newaxis]
         return M
     elif lifter == 0:
         return M
